refactor(auth): move env.ts to src/ with root compat shim
Move the environment variable definition into src/ and replace root
env.ts with a re-export for backward compatibility. Consumers that
import @xbrk/auth/env continue to work unchanged.
