+++ packages/auth/env.ts
@@ -1,30 +1,1 @@
-import { createEnv } from '@t3-oss/env-core';
-import { z } from 'zod/v4';
-
-export const env = createEnv({
-  server: {
-    NODE_ENV: z.enum(['development', 'production', 'test']).default('development'),
-
-    BETTER_AUTH_SECRET: process.env.NODE_ENV === 'production' ? z.string().min(1) : z.string().min(1).optional(),
-    BETTER_AUTH_URL: z.url().optional(),
-
-    GITHUB_CLIENT_ID: z.string().min(1).optional(),
-    GITHUB_CLIENT_SECRET: z.string().min(1).optional(),
-    GITHUB_ACCESS_TOKEN: z.string().min(1).optional(),
-
-    TWITTER_CLIENT_ID: z.string().min(1).optional(),
-    TWITTER_CLIENT_SECRET: z.string().min(1).optional(),
-
-    GOOGLE_CLIENT_ID: z.string().min(1).optional(),
-    GOOGLE_CLIENT_SECRET: z.string().min(1).optional(),
-
-    FACEBOOK_CLIENT_ID: z.string().min(1).optional(),
-    FACEBOOK_CLIENT_SECRET: z.string().min(1).optional(),
-  },
-  runtimeEnv: process.env,
-  skipValidation: !!process.env.CI || process.env.npm_lifecycle_event === 'lint',
-});
-
-export function authEnv() {
-  return env;
-}
+export * from './src/env';
